add command test for poi poly
1. refs #205

+++ hoot-core/src/main/cpp/hoot/core/filters/OrCriterion.h
@@ -52,7 +52,7 @@ public:
 
   virtual bool isSatisfied(const shared_ptr<const Element>& e) const
   {
-    for (/*size_t*/int i = 0; i < _filters.size(); i++)
+    for (size_t i = 0; i < _filters.size(); i++)
     {
       if (_filters[i]->isSatisfied(e))
       {
